requirements.txt
asgiref==3.2.10
black
certifi==2020.6.20
Django==3.1
django-common-helpers==0.9.2
django-cron==0.5.1
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.20.8
django-nine==0.2.3
django-rest-framework==0.1.0
djangorestframework==3.11.1
elasticsearch==7.9.1
elasticsearch-dsl==7.2.1
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
flake8
python-dotenv
gunicorn
dj-database-url
